Add tests for cars bloc

+++ lib/blocs/cars/bloc.dart
@@ -15,7 +15,7 @@ class CarsBloc extends Bloc<CarsEvent, CarsState> {
   static const double EMA_CUTOFF = 8;
   
   final DataRepository _dataRepository;
-  StreamSubscription _carsSubscription, _refuelingsSubscription;
+  StreamSubscription _refuelingsSubscription;
   final RefuelingsBloc _refuelingsBloc;
   List<Refueling> _refuelingsCache;
 
@@ -36,18 +36,22 @@ class CarsBloc extends Bloc<CarsEvent, CarsState> {
       yield* _mapUpdateCarToState(event);
     } else if (event is DeleteCar) {
       yield* _mapDeleteCarToState(event);
-    } else if (event is CarsUpdated) {
-      yield* _mapCarsUpdateToState(event);
     } else if (event is ExternalRefuelingsUpdated) {
       yield* _mapRefuelingsUpdatedToState(event);
     }
   }
 
   Stream<CarsState> _mapLoadCarsToState() async* {
-    _carsSubscription?.cancel();
-    _carsSubscription = _dataRepository.cars().listen(
-          (cars) => add(CarsUpdated(cars)),
-        );
+    try {
+      final cars = await _dataRepository.cars().first;
+      if (cars != null) {
+        yield CarsLoaded(cars);
+      } else {
+        yield CarsNotLoaded();
+      }
+    } catch (_) {
+      yield CarsNotLoaded();
+    }
     _refuelingsSubscription = _refuelingsBloc.listen(
       (state) {
         if (state is RefuelingsLoaded) {
@@ -58,19 +62,33 @@ class CarsBloc extends Bloc<CarsEvent, CarsState> {
   }
 
   Stream<CarsState> _mapAddCarToState(AddCar event) async* {
-    _dataRepository.addNewCar(event.car);
+    if (state is CarsLoaded) {
+      final List<Car> updatedCars = List.from((state as CarsLoaded).cars)
+        ..add(event.car);
+      yield CarsLoaded(updatedCars);
+      _dataRepository.addNewCar(event.car);
+    }
   }
 
   Stream<CarsState> _mapUpdateCarToState(UpdateCar event) async* {
-    _dataRepository.updateCar(event.updatedCar);
+    if (state is CarsLoaded) {
+      final List<Car> updatedCars = (state as CarsLoaded).cars.map((car) {
+        return car.id == event.updatedCar.id ? event.updatedCar : car;
+      }).toList();
+      yield CarsLoaded(updatedCars);
+      _dataRepository.updateCar(event.updatedCar);
+    }
   }
 
   Stream<CarsState> _mapDeleteCarToState(DeleteCar event) async* {
-    _dataRepository.deleteCar(event.car);
-  }
-
-  Stream<CarsState> _mapCarsUpdateToState(CarsUpdated event) async* {
-    yield CarsLoaded(event.cars);
+    if (state is CarsLoaded) {
+      final updatedTodos = (state as CarsLoaded)
+          .cars
+          .where((car) => car.id != event.car.id)
+          .toList();
+      yield CarsLoaded(updatedTodos);
+      _dataRepository.deleteCar(event.car);
+    }
   }
 
   Car _updateWithRefueling(Car c, Refueling r) {
@@ -80,22 +98,23 @@ class CarsBloc extends Bloc<CarsEvent, CarsState> {
       mileage = r.mileage;
       lastMileageUpdate = util.roundToDay(r.date);
     }
+    var numRefuelings = c.numRefuelings + 1;
     // Update efficiency
-    if (c.numRefuelings == 1) {
+    if (numRefuelings == 1) {
       // first refueling for this car
       averageEfficiency = r.efficiency;
     } else {
       averageEfficiency =
-          _efficiencyFilter(c.numRefuelings, c.averageEfficiency, r.efficiency);
+          _efficiencyFilter(numRefuelings, c.averageEfficiency, r.efficiency);
     }
     // Distance Rate
     var elapsedDuration = r.date.difference(c.lastMileageUpdate);
     var dist = r.mileage - c.mileage;
     var curDistRate = dist.toDouble() / elapsedDuration.inDays.toDouble();
     var distanceRate =
-        _distanceFilter(c.numRefuelings, c.distanceRate, curDistRate);
+        _distanceFilter(numRefuelings, c.distanceRate, curDistRate);
     var distanceRateHistory = c.distanceRateHistory;
-    distanceRateHistory.add(DistanceRatePoint(r.date, distanceRate));
+    distanceRateHistory.add(DistanceRatePoint(util.roundToDay(r.date), distanceRate));
 
     return c.copyWith(
       mileage: mileage,
@@ -103,38 +122,41 @@ class CarsBloc extends Bloc<CarsEvent, CarsState> {
       averageEfficiency: averageEfficiency,
       distanceRate: distanceRate,
       distanceRateHistory: distanceRateHistory,
-      numRefuelings: c.numRefuelings + 1,
+      numRefuelings: numRefuelings,
     );
   }
 
   Stream<CarsState> _mapRefuelingsUpdatedToState(ExternalRefuelingsUpdated event) async* {
     WriteBatchWrapper batch = _dataRepository.startCarWriteBatch();
     // TODO cache the cars here so that updating number of refuelings will work
+    List<Car> updatedCars = (state as CarsLoaded).cars;
     for (var r in event.refuelings) {
-      if (_refuelingsCache.contains(r)) {
+      if (_refuelingsCache?.contains(r) ?? false) {
         continue; // only do calculations for updated refuelings
       }
       
       Car cur = (state as CarsLoaded).cars.firstWhere((car) => car.name == r.carName);
       Car update = _updateWithRefueling(cur, r); 
 
+      updatedCars = (state as CarsLoaded).cars.map((car) {
+        return car.id == update.id ? update : car;
+      }).toList();
       batch.updateData(update.id, update.toEntity().toDocument());
     }
     _refuelingsCache = event.refuelings;
-    
+    yield CarsLoaded(updatedCars);
     batch.commit();
   }
 
   @override
   Future<void> close() {
-    _carsSubscription?.cancel();
     _refuelingsSubscription?.cancel();
     return super.close();
   }
 
   double _efficiencyFilter(int numRefuelings, double prev, double cur) {
     if (numRefuelings > EMA_CUTOFF) {
-      return EMA_GAIN * prev + (1 - EMA_GAIN) * cur;
+      return util.roundToPrecision(EMA_GAIN * prev + (1 - EMA_GAIN) * cur, 3);
     } else {
       double fac1 = (numRefuelings - 1) / numRefuelings;
       double fac2 = 1 / numRefuelings;
